Removed an unnecessary new payload call, moved Mappers to the same place, added some more logs
