+++ app/build.gradle
@@ -10,6 +10,7 @@ dependencies {
   implementation(project(":storage"))
   implementation(project(":serialization"))
   implementation(project(":jvm-libs:extensions"))
+  implementation(project(":jvm-libs:mappers"))
 
   implementation "info.picocli:picocli"
   implementation "com.sksamuel.hoplite:hoplite-core"
@@ -21,7 +22,6 @@ dependencies {
   testImplementation(project(":jvm-libs:test-utils"))
   testImplementation(testFixtures(project(":core")))
   testImplementation(testFixtures(project(":config")))
-  testImplementation("org.hyperledger.besu.internal:core")
   testImplementation "org.hyperledger.besu.internal:api"
   testImplementation(group: 'org.hyperledger.besu.internal', name: 'core', classifier: 'test-support')
