+++ app/src/test/resources/log4j2-test.xml
@@ -14,7 +14,7 @@
             level="TRACE"/>
 <!--    <Logger name="org.hyperledger.besu.consensus.merge.blockcreation.MergeCoordinator"-->
 <!--            level="TRACE"/>-->
-    <Logger name="org.hyperledger.besu" level="INFO"/>
+    <Logger name="org.hyperledger.besu" level="DEBUG"/>
     <Logger name="org.hyperledger.besu.consensus.qbft" level="TRACE"/>
     <Root level="INFO" additivity="false">
       <appender-ref ref="console"/>
