+++ e2e/build.gradle
@@ -9,6 +9,7 @@ dependencies {
   implementation project(":app")
   implementation project(":consensus")
   implementation project(":config")
+  implementation project(":jvm-libs:mappers")
 
   implementation "tech.pegasys.teku.internal:executionclient"
   implementation "tech.pegasys.teku.internal:time"
