+++ settings.gradle
@@ -14,3 +14,4 @@ include 'storage'
 include 'serialization'
 include 'config'
 include 'jvm-libs:test-utils'
+include 'jvm-libs:mappers'
